test(coscrad-frontend): remove skipped test
