+++ apps/coscrad-frontend/src/components/dynamicViews/DataPresenters/DefaultPresenter.tsx
@@ -1,0 +1,1 @@
+export const DefaultPresenter = (data: unknown) => <div>{JSON.stringify(data)}</div>;
