+++ apps/coscrad-frontend/src/components/dynamicViews/DataPresenters/index.tsx
@@ -1,0 +1,1 @@
+export * from './buildPresenterForPropertyFromDataType';
