+++ apps/coscrad-frontend/src/components/dynamicViews/index.tsx
@@ -1,0 +1,1 @@
+export * from './buildViewModelComponentFromCoscradDataDefinition';
