+++ apps/coscrad-frontend/tsconfig.spec.json
@@ -16,7 +16,7 @@
         "**/*.spec.jsx",
         "**/*.d.ts",
         "jest.config.ts"
-    ],
+,    ],
     "files": [
         "../../node_modules/@nrwl/react/typings/cssmodule.d.ts",
         "../../node_modules/@nrwl/react/typings/image.d.ts"
