+++ libs/data-types/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/enums';
 export * from './lib/fuzz';
 export * from './lib/types';
 export * from './lib/utilities';
+export * from './test/widget';
