+++ libs/validation/src/index.ts
@@ -23,4 +23,5 @@ export {
 } from 'class-validator';
 export * from './lib/decorators';
 export * from './lib/interfaces';
+export * from './lib/validationFunctions';
 export { buildSimpleValidationFunction, DiscriminatedUnionValidator };
