+++ libs/validation/src/lib/validationFunctions/index.ts
@@ -1,0 +1,2 @@
+export { isBoolean } from 'class-validator';
+export * from './isNullOrUndefined';
