+++ libs/validation/src/lib/validationFunctions/isNullOrUndefined.ts
@@ -1,0 +1,2 @@
+export const isNullOrUndefined = (input: unknown): input is null | undefined =>
+    input === null || typeof input === 'undefined';
